- mpved all software renderer specific code for the portals into the render sources.
